Implement q_ascend and q_descend
Add q_ascend to remove every node which has a node with a
strictly less value anywhere to the right side of it.
Add q_descend to remove every node which has a node with a
strictly greater value anywhere to the right side of it.

Firstly, reverse whole queue.
Secondly, check if queue is in ascending order or descending order.
Then delete the nodes that against the order.
Finally, reverse whole queue again.
Return the size of new queue.

Change-Id: I443dbc326ea3bd7810989767c7dc55cd4b64990f