Add q_sort and two sub-function of it
Add q_sort to sort elements of queue in ascending or descending order.
There are two step which are partition and merge.

In partition, head pointer points to the first element of queue and
tail pointer points to the next pointer of the last element.
Therefore, the element of last pointer does not belong to the target.
I separate target queue into two part which are left and right.
Then recursively until target queue only have one member.
Finally, use merge function to trace back.

In merge, we sort the queue.
The head of left queue is left pointer and the tail is right pointer.
The head of right queue is right pointer and the tail is tail pointer.
Therefore, if left equal to right, mean left queue is empty.
If right equal to tail, mean right queue is empty, then we return.
Each iteration, we compare each element in both queue and sort.

However, using make check will appear error.
Not stable sort. The duplicate strings are not in the same order.

Change-Id: I4fd16b8f2b398aafb01d4090b8b4fde9c792e79c